+++ src/helpers/firstValues.js
@@ -1,0 +1,16 @@
+import { multipartType } from "../plugins/multipart.js";
+import { querystringType } from "../plugins/querystring.js";
+
+export { firstValues };
+
+const firstValues = (form, fields, exceptions = []) => {
+    if (form.type !== querystringType && form.type !== multipartType) {
+        return fields;
+    }
+    return Object.fromEntries(Object.entries(fields).map(([key, value]) => {
+        if (exceptions.includes(key)) {
+            return [key, value];
+        }
+        return [key, value[0]];    
+    }));
+};