+++ src/plugins/octetstream.js
@@ -3,6 +3,7 @@
 
 import OctetStreamParser from '../parsers/OctetStream.js';
 
+export const octetStreamType = 'octet-stream';
 // the `options` is also available through the `options` / `formidable.options`
 export default function plugin(formidable, options) {
   // the `this` context is always formidable, as the first argument of a plugin
@@ -22,7 +23,7 @@ export default function plugin(formidable, options) {
 // of the passed `options` (second) param, because when you decide
 // to test the plugin you can pass custom `this` context to it (and so `this.options`)
 function init(_self, _opts) {
-  this.type = 'octet-stream';
+  this.type = octetStreamType;
   const originalFilename = this.headers['x-file-name'];
   const mimetype = this.headers['content-type'];