+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Unreleased 3.0
 
+ * feat: remove options.multiples ([730](https://github.com/node-formidable/formidable/pull/730))
+ * use modern URLSearchParams https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams internally
+ * files and fields values are always arrays
+ * remove unused qs and querystring dependency
  * feat: Use ES modules ([727](https://github.com/node-formidable/formidable/pull/727))
  * options.enabledPlugins must contain the plugin themselves instead of the plugins names